+++ paymentsheet/src/test/java/com/stripe/android/paymentsheet/addresselement/AddressLauncherFixtures.kt
@@ -1,5 +1,5 @@
 package com.stripe.android.paymentsheet.addresselement
 
 internal object AddressLauncherFixtures {
-    internal val BASIC_CONFIG = AddressLauncher.Configuration()
+    internal val BASIC_CONFIG = AddressLauncher.Configuration.Builder().build()
 }